[breadboard-cli] Loader abstract class had a lot of specific ts code, move into ts-loader
